Extensions
It is worthy of note that only a *single* extension (`stack-depth')
implements functionality which is unavailable in the original false
specification.  As such, these extensions do not add any additional
richness to the language.

Assert that stack top is not false
 * ∫
 * compact expression of ``0=[}]?``
 * useful for writing test cases

Less-than
 * <
 * compact expression of ``\>``

Not-equal-to
 * ≠
 * compact expression of ``=~``

Less-than-or-equal-to
 * ≤
 * compact expression of ``>~``

Greater-than-or-equal-to
 * ≥
 * compact expression of ``\>~``

Divide
 * ÷  ( x y , rem quot )
 * compact expression of ``1ø1ø1ø1ø/*-@@/``
 * ``÷%`` provides the remainder only like ``1ø1ø/*-``

Bitwise exclusive or
 * ⊻
 * compact expression of ``$@$@|@@&~&``
 * U+22BB is not easily typed, but is a useful operator

Binary shift left
 * «

Binary shift right
 * »

Stack-depth
 * §

(Forth) over
 * £  ( a b , a b a )
 * compact expression of ``\$@\``

(Forth) nip
 * ‰
 * compact expression of ``\%``

(Forth) tuck (Java) dup_x1
 * €  ( a b , b a b )
 * compact expression of ``$@\``

(Forth) 2dup (Java) dup2
 * Ø  ( a b , a b a b )
 * compact expression of ``1ø1ø``

If-then-else
 * ¿
 * compact expression of ``$[true]?~[false]?``

Signed-off-by: Ian Ray <ianjray@me.com>